extract/postgres/requirements.txt
fire
mypy
pandas
psycopg2-binary
pyyaml
snowflake-sqlalchemy
